Rename the file bookSlice.js to booksSlice.js
