+++ src/redux/store.js
@@ -1,10 +1,10 @@
 import { configureStore } from '@reduxjs/toolkit';
-import bookSlice from './books/bookSlice';
+import booksSlice from './books/booksSlice';
 import categoriesSlice from './categories/categoriesSlice';
 
 const store = configureStore({
   reducer: {
-    books: bookSlice,
+    books: booksSlice,
     categories: categoriesSlice,
   },
 });
